src/loop_closure/requirements.txt
appnope==0.1.3
asttokens==2.0.8
av==9.2.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fonttools==4.37.4
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
stack-data==0.5.1
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
traitlets==5.4.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
